+++ sop-core/src/Data/SOP/Constraint.hs
@@ -112,11 +112,12 @@ type SListI2 = All SListI
 -- a type-level list argument that the function is quantified
 -- over.
 --
--- This is defined in terms of 'All'. A singleton list provides
--- a witness for a type-level list where the elements adhere
--- to no additional constraints.
+-- Since 0.4.0.0, this is now defined in terms of 'All'.
+-- A singleton list provides a witness for a type-level list
+-- where the elements need not satisfy any additional
+-- constraints.
 --
--- @since 0.2
+-- @since 0.4.0.0
 --
 type SListI = All Top
 
@@ -141,9 +142,14 @@ type SListI = All Top
 -- means that 'f' can assume that all elements of the sum
 -- of product satisfy 'Eq'.
 --
+-- Since 0.4.0.0, this is merely a synonym for
+-- 'All (All c)'.
+--
+-- @since 0.4.0.0
+--
 type All2 c = All (All c)
 
--- | Require a constraint for pointwise for every pair of
+-- | Require a constraint pointwise for every pair of
 -- elements from two lists.
 --
 -- /Example:/ The constraint
@@ -205,15 +211,15 @@ type family Head (xs :: [a]) :: a where
 type family Tail (xs :: [a]) :: [a] where
   Tail (x ': xs) = xs
 
--- | The constraint @LiftedCoercible f g x y@ is equivalent
--- to @Coercible (f x) (g y)@.
+-- | The constraint @'LiftedCoercible' f g x y@ is equivalent
+-- to @'Data.Coerce.Coercible' (f x) (g y)@.
 --
 -- @since 0.3.1.0
 --
 class Coercible (f x) (g y) => LiftedCoercible f g x y
 instance Coercible (f x) (g y) => LiftedCoercible f g x y
 
--- | Require a constraint for pointwise for every pair of
+-- | Require a constraint pointwise for every pair of
 -- elements from two lists of lists.
 --
 --
@@ -223,13 +229,13 @@ instance (AllZipF (AllZip f) xss yss, SListI xss, SListI yss, SameShapeAs xss ys
 -- | Composition of constraints.
 --
 -- Note that the result of the composition must be a constraint,
--- and therefore, in @f ':.' g@, the kind of @f@ is @k -> 'Constraint'@.
--- The kind of @g@, however, is @l -> k@ and can thus be an normal
+-- and therefore, in @'Compose' f g@, the kind of @f@ is @k -> 'Constraint'@.
+-- The kind of @g@, however, is @l -> k@ and can thus be a normal
 -- type constructor.
 --
--- A typical use case is in connection with 'All' on an 'NP' or an
--- 'NS'. For example, in order to denote that all elements on an
--- @'NP' f xs@ satisfy 'Show', we can say @'All' ('Show' :. f) xs@.
+-- A typical use case is in connection with 'All' on an 'Data.SOP.NP' or an
+-- 'Data.SOP.NS'. For example, in order to denote that all elements on an
+-- @'Data.SOP.NP' f xs@ satisfy 'Show', we can say @'All' ('Compose' 'Show' f) xs@.
 --
 -- @since 0.2
 --
@@ -272,4 +278,3 @@ type family AllZipN (h :: (k -> Type) -> (l -> Type)) (c :: k1 -> k2 -> Constrai
 -- on whether the argument is indexed by a list or a list of lists.
 --
 type family SListIN (h :: (k -> Type) -> (l -> Type)) :: l -> Constraint
-